% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darwinator-package.R
\name{darwinator-package}
\alias{darwinator-package}
\alias{darwinator}
\title{darwinator}
\description{
This package provides tools for working with sampling event
datasets in DwcA (Darwin Core Archive) format
}
\keyword{package}
