% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_event_data.R
\name{sampling_event_data}
\alias{sampling_event_data}
\title{Sampling Event Data Download from IPT}
\usage{
sampling_event_data(key)
}
\arguments{
\item{key}{character string with sampling event dataset identifier from GBIF}
}
\value{
A list with slots for metadata (`meta`), for the DwcA tables (`dwca`) and all the data from the various sampling event data tables joined into one data frame (`data`)
}
\description{
This function downloads DwC-A directly from the IPT. The GBIF API does not
yet throw back empty events and potential hierarchical
structures which may be essential for inferring sampling effort. This is under
implementation at GBIF and we hope that this function will soon be redundant.
}
\examples{
\dontrun{
sed <- sampling_event_data("78360224-5493-45fd-a9a0-c336557f09c3")
}
}
